changing longs to decimal to avoid overflow on high values

+++ SimpleHashDotCom/SimpleHashDotCom.Sdk/Models/Nft/AudioProperties.cs
@@ -4,15 +4,15 @@ namespace SimpleHashDotCom.Sdk.Models.Nft;
 
 public class AudioProperties
 {
-	[JsonPropertyName("duration")]
-	public decimal? Duration { get; set; }
+    [JsonPropertyName("duration")]
+    public decimal? Duration { get; set; }
 
-	[JsonPropertyName("audio_coding")]
-	public string? AudioCoding { get; set; }
+    [JsonPropertyName("audio_coding")]
+    public string? AudioCoding { get; set; }
 
-	[JsonPropertyName("size")]
-	public long? Size { get; set; }
+    [JsonPropertyName("size")]
+    public decimal? Size { get; set; }
 
-	[JsonPropertyName("mime_type")]
-	public string? MimeType { get; set; }
+    [JsonPropertyName("mime_type")]
+    public string? MimeType { get; set; }
 }
