+++ SimpleHashDotCom/SimpleHashDotCom.Sdk/Models/Nft/ModelProperties.cs
@@ -4,9 +4,9 @@ namespace SimpleHashDotCom.Sdk.Models.Nft;
 
 public class ModelProperties
 {
-	[JsonPropertyName("size")]
-	public long? Size { get; set; }
+    [JsonPropertyName("size")]
+    public decimal? Size { get; set; }
 
-	[JsonPropertyName("mime_type")]
-	public string? MimeType { get; set; }
+    [JsonPropertyName("mime_type")]
+    public string? MimeType { get; set; }
 }
